pyproject.toml
[project]
name = "slackbotparty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.12",
    "python-dotenv>=1.0.1",
    "slack-bolt>=1.22.0",
    "slack-sdk>=3.34.0",
]
